Refactored code and moved methods to new classes or more appropriate classes; simplified code in a few places. Still need to write new tests.
