+++ test/unit/php/includes/core/classes/class-test-event-setup.php
@@ -1,0 +1,96 @@
+<?php
+/**
+ * Class handles unit tests for GatherPress\Core\Event_Setup.
+ *
+ * @package GatherPress\Core
+ * @since 1.0.0
+ */
+
+namespace GatherPress\Tests\Core;
+
+use DateTime;
+use GatherPress\Core\Event;
+use GatherPress\Core\Event_Setup;
+use GatherPress\Core\Topic;
+use GatherPress\Core\Venue;
+use PMC\Unit_Test\Base;
+
+/**
+ * Class Test_Event_Query.
+ *
+ * @coversDefaultClass \GatherPress\Core\Event_Setup
+ */
+class Test_Event_Setup extends Base {
+	/**
+	 * Coverage for setup_hooks.
+	 *
+	 * @covers ::__construct
+	 * @covers ::setup_hooks
+	 *
+	 * @return void
+	 */
+	public function test_setup_hooks(): void {
+		$instance = Event_Setup::get_instance();
+		$hooks    = array(
+			array(
+				'type'     => 'action',
+				'name'     => 'delete_post',
+				'priority' => 10,
+				'callback' => array( $instance, 'delete_event' ),
+			),
+			array(
+				'type'     => 'filter',
+				'name'     => sprintf( 'manage_%s_posts_columns', Event::POST_TYPE ),
+				'priority' => 10,
+				'callback' => array( $instance, 'set_custom_columns' ),
+			),
+			array(
+				'type'     => 'filter',
+				'name'     => sprintf( 'manage_edit-%s_sortable_columns', Event::POST_TYPE ),
+				'priority' => 10,
+				'callback' => array( $instance, 'sortable_columns' ),
+			),
+			array(
+				'type'     => 'filter',
+				'name'     => 'get_the_date',
+				'priority' => 10,
+				'callback' => array( $instance, 'get_the_event_date' ),
+			),
+			array(
+				'type'     => 'filter',
+				'name'     => 'the_time',
+				'priority' => 10,
+				'callback' => array( $instance, 'get_the_event_date' ),
+			),
+			array(
+				'type'     => 'filter',
+				'name'     => 'display_post_states',
+				'priority' => 10,
+				'callback' => array( $instance, 'set_event_archive_labels' ),
+			),
+		);
+
+		$this->assert_hooks( $hooks, $instance );
+	}
+	/**
+	 * Coverage for sortable_columns method.
+	 *
+	 * @covers ::sortable_columns
+	 *
+	 * @return void
+	 */
+	public function test_sortable_columns(): void {
+		$instance = Event_Setup::get_instance();
+		$default  = array( 'unit' => 'test' );
+		$expects  = array(
+			'unit'     => 'test',
+			'datetime' => 'datetime',
+		);
+
+		$this->assertSame(
+			$expects,
+			$instance->sortable_columns( $default ),
+			'Failed to assert correct sortable columns.'
+		);
+	}
+}
