+++ test/unit/php/includes/core/classes/class-test-event.php
@@ -45,63 +45,6 @@ class Test_Event extends Base {
 	}
 
 	/**
-	 * Coverage for get_post_type_registration_args method.
-	 *
-	 * @covers ::get_post_type_registration_args
-	 *
-	 * @return void
-	 */
-	public function test_get_post_type_registration_args(): void {
-		$args = Event::get_post_type_registration_args();
-
-		$this->assertIsArray( $args['labels'], 'Failed to assert that labels are an array.' );
-		$this->assertTrue( $args['show_in_rest'], 'Failed to assert that show_in_rest is true.' );
-		$this->assertTrue( $args['public'], 'Failed to assert that public is true.' );
-		$this->assertSame( 'dashicons-nametag', $args['menu_icon'], 'Failed to assert that menu_icon is nametag.' );
-		$this->assertSame( 'event', $args['rewrite']['slug'], 'Failed to assert that slug is events.' );
-	}
-
-	/**
-	 * Coverage for get_post_meta_registration_args method.
-	 *
-	 * @covers ::get_post_meta_registration_args
-	 *
-	 * @return void
-	 */
-	public function test_get_post_meta_registration_args(): void {
-		$args = Event::get_post_meta_registration_args();
-
-		$this->assertIsArray(
-			$args['online_event_link'],
-			'Failed to assert that online_event_link is an array.'
-		);
-		$this->assertIsArray(
-			$args['enable_anonymous_rsvp'],
-			'Failed to assert that online_event_link is an array.'
-		);
-
-		$this->mock->user( 'subscriber' );
-		$this->assertFalse(
-			$args['online_event_link']['auth_callback'](),
-			'Failed to assert that user cannot edit posts.'
-		);
-		$this->assertFalse(
-			$args['enable_anonymous_rsvp']['auth_callback'](),
-			'Failed to assert that user cannot edit posts.'
-		);
-
-		$this->mock->user( 'admin' );
-		$this->assertTrue(
-			$args['online_event_link']['auth_callback'](),
-			'Failed to assert that user can edit posts.'
-		);
-		$this->assertTrue(
-			$args['enable_anonymous_rsvp']['auth_callback'](),
-			'Failed to assert that user can edit posts.'
-		);
-	}
-
-	/**
 	 * Data provider for get_display_datetime test.
 	 *
 	 * @return array
@@ -328,84 +271,6 @@ class Test_Event extends Base {
 			Utility::invoke_hidden_method( $event, 'get_gmt_datetime', array( 'unit-test', $timezone ) ),
 			'Failed to assert that gmt datetime matches.'
 		);
-	}
-
-	/**
-	 * Data provider for maybe_convert_offset test.
-	 *
-	 * @return array
-	 */
-	public function data_maybe_convert_offset(): array {
-		return array(
-			array(
-				'America/New_York',
-				'America/New_York',
-			),
-			array(
-				'UTC',
-				'UTC',
-			),
-			array(
-				'UTC+9.5',
-				'+09:30',
-			),
-			array(
-				'UTC-7.25',
-				'-07:15',
-			),
-			array(
-				'UTC-5.75',
-				'-05:45',
-			),
-			array(
-				'UTC+1',
-				'+01:00',
-			),
-		);
-	}
-
-	/**
-	 * Coverage for maybe_convert_offset method.
-	 *
-	 * @dataProvider data_maybe_convert_offset
-	 *
-	 * @covers ::maybe_convert_offset
-	 *
-	 * @param string $input   Value to pass to method.
-	 * @param string $expects Expected response.
-	 *
-	 * @return void
-	 */
-	public function test_maybe_convert_offset( $input, $expects ): void {
-		$this->assertSame(
-			$expects,
-			Event::maybe_convert_offset( $input ),
-			'Failed to assert that conversion matches.'
-		);
-	}
-
-	/**
-	 * Coverage for list_identifiers method.
-	 *
-	 * @covers ::list_identifiers
-	 *
-	 * @return void
-	 */
-	public function test_list_identifiers(): void {
-		$list      = Event::list_identifiers();
-		$timezones = array(
-			'America/Belem',
-			'Asia/Chita',
-			'Europe/Vilnius',
-			'UTC',
-			'-12:00',
-			'-00:30',
-			'+09:30',
-			'+13:45',
-		);
-		foreach ( $timezones as $timezone ) {
-			$this->assertContains( $timezone, $list, 'Failed to assert timezone is in list.' );
-		}
 	}
 
 	/**
